% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_genes}
\alias{BRCA_genes}
\title{BRCA genes}
\format{BRCA_genes: A SummarizedExperiment object with
4819 BRCA related genes (including lncRNAs and mRNAs).}
\description{
BRCA genes
}
\details{
The BRCA related lncRNAs are from LncRNADisease v2.0, Lnc2Cancer v2.0 
and MNDR v2.0. The BRCA related mRNAs are from DisGeNET v5.0 and COSMIC v86.
}
\references{
Bao Z, Yang Z, Huang Z, Zhou Y, Cui Q, Dong D. (2019) 
"LncRNADisease 2.0: an updated database of long non-coding RNA-associated diseases".
Nucleic Acids Res., 47(D1):D1034-D1037.

Cui T, Zhang L, Huang Y, Yi Y, Tan P, Zhao Y, Hu Y, Xu L, Li E, Wang D. 
(2018) "MNDR v2.0: an updated resource of ncRNA-disease 
associa-tions in mammals". Nucleic Acids Res., 46, D371-D374.

Gao Y, Wang P, Wang Y, Ma X, Zhi H, Zhou D, Li X, Fang Y, Shen W, Xu Y, 
Shang S, Wang L, Wang L, Ning S, Li X. (2019) "Lnc2Cancer v2.0: updated database of experimentally 
supported long non-coding RNAs in human cancers". Nucleic Acids Res., 47, D1028-D1033.

Forbes SA, Beare D, Boutselakis H, Bamford S, Bindal N, Tate J, Cole CG, 
Ward S, Dawson E, Ponting L, Stefancsik R, Harsha B, Kok CY, Jia M, Jubb H, Sondka Z, 
Thompson S, De T, Campbell PJ. (2017) "COSMIC: somatic cancer genetics at 
high-resolution". Nucleic Acids Res., 45, D777-D783

Pinero J, Bravo A, Queralt-Rosinach N, Gutierrez-Sacristan A, Deu-Pons J, 
Centeno E, Garcia-Garcia J, Sanz F, Furlong LI. 
(2017) "DisGeNET: a comprehensive platform integrating 
infor-mation on human disease-associated genes and variants". 
Nucleic Acids Res., 45, D833-D839.
}
