% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRSM.R
\name{data_tidy}
\alias{data_tidy}
\title{data_tidy}
\usage{
data_tidy(Exp)
}
\arguments{
\item{Exp}{Input expression data: rows are sample names and columns are gene names.}
}
\value{
Output expression data matrix.
}
\description{
Tidy the input expression data matrix to reserve "-" symbols of gene names
}
\examples{
ceRExpcsv <- system.file("extdata","ceRExp.csv",package="miRSM")
ceRExp <- read.csv(ceRExpcsv, header=FALSE, sep=",")
ceRExp <- data_tidy(ceRExp)

}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng_Zhang3})
}
