% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRExp}
\alias{miRExp}
\title{miRNA expression data.}
\format{A SummarizedExperiment object with 144 samples (rows) and 226 miRNAs (columns).}
\usage{
miRExp
}
\description{
A dataset containing the miRNA expressopn levels for 144 TCGA samples (72 cancer samples
vs 72 normal samples, https://cancergenome.nih.gov/)
}
\keyword{datasets}
