% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRSM.R
\name{module_biclust}
\alias{module_biclust}
\title{module_biclust}
\usage{
module_biclust(
  ceRExp,
  mRExp = NULL,
  BCmethod = "fabia",
  num.modules = 10,
  num.ModuleceRs = 2,
  num.ModulemRs = 2
)
}
\arguments{
\item{ceRExp}{A SummarizedExperiment object. ceRNA expression data: 
rows are samples and columns are ceRNAs.}

\item{mRExp}{NULL (default) or a SummarizedExperiment object. mRNA expression data: 
rows are samples and columns are mRNAs.}

\item{BCmethod}{Specification of the biclustering method, 
including 'fabia' (default), 'fabiap', 
'fabias', 'mfsc', 'nmfdiv', 'nmfeu', 'nmfsc', 'FLOC' and 'isa'.}

\item{num.modules}{The number of modules to be identified. For the 
'isa' method, no need to set the parameter.}

\item{num.ModuleceRs}{The minimum number of ceRNAs in each module.}

\item{num.ModulemRs}{The minimum number of mRNAs in each module.}
}
\value{
GeneSetCollection object: a list of module genes.
}
\description{
Identification of gene modules from matched ceRNA and mRNA 
expression data or single gene expression data using a series  
of biclustering packages, including fabia,  
BicARE and isa2
}
\examples{
data(BRCASampleData)
modulegenes_biclust <- module_biclust(ceRExp[, seq_len(30)],
    mRExp[, seq_len(30)])

}
\references{
Hochreiter S, Bodenhofer U, Heusel M, Mayr A, 
Mitterecker A, Kasim A, Khamiakova T, Van Sanden S, Lin D, 
Talloen W, Bijnens L, G\'{o}hlmann HW, Shkedy Z, Clevert DA. 
FABIA: factor analysis for bicluster acquisition. 
Bioinformatics. 2010, 26(12):1520-7.

Yang J, Wang H, Wang W, Yu, PS. An improved 
biclustering method for analyzing gene expression. 
Int J Artif Intell Tools. 2005, 14(5): 771-789.

Bergmann S, Ihmels J, Barkai N. Iterative 
signature algorithm for the analysis of large-scale gene 
expression data. Phys Rev E Stat Nonlin Soft Matter Phys. 
2003, 67(3 Pt 1):031902.

Sill M, Kaiser S, Benner A, Kopp-Schneider A. 
Robust biclustering by sparse singular value decomposition 
incorporating stability selection. Bioinformatics. 2011, 
27(15):2089-97.
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng-Zhang-2})
}
