% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRSM.R
\name{module_FA}
\alias{module_FA}
\title{module_FA}
\usage{
module_FA(Modulelist, GOont = "BP", Diseaseont = "DO",
  KEGGorganism = "hsa", Reactomeorganism = "human",
  OrgDb = "org.Hs.eg.db", padjustvaluecutoff = 0.05,
  padjustedmethod = "BH", Analysis.type = c("FEA", "DEA"))
}
\arguments{
\item{Modulelist}{List object: a list of miRNA sponge modules.}

\item{GOont}{One of "MF", "BP", and "CC" subontologies.}

\item{Diseaseont}{One of "DO", and "DOLite" subontologies.}

\item{KEGGorganism}{Organism, supported organism listed in http://www.genome.jp/kegg/catalog/org_list.html.}

\item{Reactomeorganism}{Organism, one of "human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly".}

\item{OrgDb}{OrgDb}

\item{padjustvaluecutoff}{A cutoff value of adjusted p-values.}

\item{padjustedmethod}{Adjusted method of p-values, can select one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{Analysis.type}{The type of functional analysis selected, including "FEA" (functional enrichment analysis) and "DEA" (disease enrichment analysis).}
}
\value{
List object: a list of enrichment analysis results.
}
\description{
Functional analysis of miRNA sponge modules, including functional enrichment and disease enrichment analysis
}
\examples{
\dontrun{
data(ceRExp)
data(mRExp)
data(miRExp)
data(miRTarget)
modulegenes_WGCNA <- module_WGCNA(ceRExp, mRExp)
# Identify miRNA sponge modules using cannonical correlation (CC)
miRSM_WGCNA_CC <- miRSM(miRExp, ceRExp, mRExp, miRTarget,
                        modulegenes_WGCNA, nperms = 10, method = "CC")
miRSM_WGCNA_CC_genes <- miRSM_WGCNA_CC[[2]]
miRSM_WGCNA_CC_FEA <- module_FA(miRSM_WGCNA_CC_genes, Analysis.type = "FEA")
miRSM_WGCNA_CC_DEA <- module_FA(miRSM_WGCNA_CC_genes, Analysis.type = "DEA")
}

}
\references{
Zhang J (2017). miRsponge: Identification and analysis of miRNA sponge interaction networks and modules. R package version 1.2.0, (\url{https://github.com/zhangjunpeng411/miRsponge}).
}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng_Zhang3})
}
