% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRSM.R
\name{share_miRs}
\alias{share_miRs}
\title{share_miRs}
\usage{
share_miRs(miRExp, ceRExp, mRExp, miRTarget, Modulelist)
}
\arguments{
\item{miRExp}{A SummarizedExperiment object. miRNA expression data: 
rows are samples and columns are miRNAs.}

\item{ceRExp}{A SummarizedExperiment object. ceRNA expression data: 
rows are samples and columns are ceRNAs.}

\item{mRExp}{A SummarizedExperiment object. mRNA expression data: 
rows are samples and columns are mRNAs.}

\item{miRTarget}{A SummarizedExperiment object. Putative 
miRNA-target binding information.}

\item{Modulelist}{List object: a list of the identified miRNA sponge modules.}
}
\value{
List object: a list of common miRNAs of each miRNA sponge module.
}
\description{
Extract common miRNAs of each miRNA sponge module
}
\examples{
data(BRCASampleData)
modulegenes_WGCNA <- module_WGCNA(ceRExp, mRExp)
# Identify miRNA sponge modules using canonical correlation (CC)
miRSM_WGCNA_CC <- miRSM(miRExp, ceRExp, mRExp, miRTarget,
                        modulegenes_WGCNA, nperms = 10, method = 'CC')
miRSM_WGCNA_CC_genes <- miRSM_WGCNA_CC[[2]]
miRSM_WGCNA_share_miRs <-  share_miRs(miRExp, ceRExp, mRExp, miRTarget, miRSM_WGCNA_CC_genes)

}
\author{
Junpeng Zhang (\url{https://www.researchgate.net/profile/Junpeng_Zhang3})
}
