% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRExp}
\alias{miRExp}
\title{miRNA expression data}
\format{miRExp: A SummarizedExperiment object with 72 BRCA and 72 normal
samples (rows) and 226 miRNAs (columns).}
\description{
miRNA expression data
}
\details{
The matched breast invasive carcinoma (BRCA) miRNA, lncRNA 
and mRNA expression data is obtained from TCGA 
(http://cancergenome.nih.gov/). The data focuses on 
72 individuals for which the complete sets of 
tumor and matched normal (i.e., normal tissue taken from the 
same patient) profiles are available. 
A miRNA which has missing values in more than 
10% of the samples is removed. The remaining missing values
are imputed using the k-nearest neighbours (KNN) algorithm
from the impute R package. We use the limma R package
to infer differentially expressed miRNAs, ceRNAs and mRNAs
between tumour and normal samples. After the analysis, 
we select top 226 miRNAs which are differentially expressed 
at a significant level (adjusted p-value < 1E-02,  
adjusted by Benjamini & Hochberg method).
}
